pyproject.toml
[tool.poetry]
name = "gestor-estoque"
version = "0.1.0"
description = ""
authors = ["vitorLourival22 <vitorLou2204@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-crispy-forms = "^2.2"
crispy-bootstrap5 = "^2024.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I','F','E','W','PL','PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
run = 'python manage.py runserver'
lint = 'ruff check , && ruff check , --diff'
format = 'ruff check , --fix && ruff format ,'